requirements.txt
streamlit
langchain
langchain-community
langchain-experimental
langchain-openai
openai
pandas
openpyxl
python-dotenv